The value of the price variable in the getDataPoint function was changed so that it gave the average for the bid and asking price of a stock.
The getRatio function was fixed so that it gave a ratio between the
prices of the two stocks. Additionally, an if statement was created to
prevent the possibility of a division by zero when performing the ratio.

Inside Main, a dictionary was created to store the prices of the two
stocks. Inside the nested for loop, the dictionary is added with these
values. Lastly, the final print statement has been changed so that it
grabs and prints the prices for ABC and DEF.